+++ Model/FormRecord/Attribute.php
@@ -5,12 +5,19 @@
  */
 namespace Alekseon\CustomFormsBuilder\Model\FormRecord;
 
+use Alekseon\CustomFormsBuilder\Model\FieldOptionSources;
+
 /**
  * Class Attribute
  * @package Alekseon\CustomFormsBuilder\Model\FormRecord
  */
 class Attribute extends \Alekseon\AlekseonEav\Model\Attribute
 {
+    /**
+     * @var \Alekseon\CustomFormsBuilder\Model\FieldOptionSources
+     */
+    protected $fieldOptionSources;
+
     /**
      * Attribute constructor.
      * @param \Magento\Framework\Model\Context $context
@@ -19,6 +26,7 @@ class Attribute extends \Alekseon\AlekseonEav\Model\Attribute
      * @param \Alekseon\AlekseonEav\Model\Attribute\InputValidatorRepository $inputValidatorRepository
      * @param \Alekseon\CustomFormsBuilder\Model\ResourceModel\FormRecord\Attribute $resource
      * @param \Alekseon\CustomFormsBuilder\Model\ResourceModel\FormRecord\Attribute\Collection $resourceCollection
+     * @param \Alekseon\CustomFormsBuilder\Model\FieldOptionSources $fieldOptionSources
      */
     public function __construct(
         \Magento\Framework\Model\Context $context,
@@ -26,10 +34,39 @@ class Attribute extends \Alekseon\AlekseonEav\Model\Attribute
         \Alekseon\AlekseonEav\Model\Attribute\InputTypeRepository $inputTypeRepository,
         \Alekseon\AlekseonEav\Model\Attribute\InputValidatorRepository $inputValidatorRepository,
         \Alekseon\CustomFormsBuilder\Model\ResourceModel\FormRecord\Attribute $resource,
-        \Alekseon\CustomFormsBuilder\Model\ResourceModel\FormRecord\Attribute\Collection $resourceCollection
+        \Alekseon\CustomFormsBuilder\Model\ResourceModel\FormRecord\Attribute\Collection $resourceCollection,
+        FieldOptionSources $fieldOptionSources
     ) {
+        $this->fieldOptionSources = $fieldOptionSources;
         parent::__construct(
             $context, $registry, $inputTypeRepository, $inputValidatorRepository, $resource, $resourceCollection
         );
+    }
+
+    /**
+     * @return \Alekseon\AlekseonEav\Model\Attribute
+     */
+    public function beforeSave()
+    {
+        if ($optionSourceCode = $this->getOptionSourceCode()) {
+
+            $optionSource = $this->fieldOptionSources->getOptionSourceByCode($optionSourceCode);
+            $backendType = $this->getInputTypeModel()->getDefaultBackendType();
+
+            if ($optionSource) {
+                $this->setSourceModel($optionSource->getSourceModel());
+                if ($optionSource->getBackendType()) {
+                    $backendType = $optionSource->getBackendType();
+                }
+                $this->setBackendType($backendType);
+            }
+
+            if ($optionSourceCode == FieldOptionSources::DEFAULT_ATTRIBUTE_SOURCE_VALUE) {
+                $this->setSourceModel(null);
+                $this->setBackendType($backendType);
+            }
+        }
+
+        return parent::beforeSave();
     }
 }