+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 ### Added
 
+## [Unreleased]
+### Added
+- option sources for select attributes
+
 ## [102.0.0] - 2020-10-15
 ### Added
 - Added compatibility with Magento 2.4.0